requirements.txt
lxml>=5.2.1
click>=8.1.7
pandas>=2.2.2
pyxlsb>=1.0.10
pyyaml>=6.0.2
matplotlib>=3.8